WIP abstract MetricsLayout
