+++ runtime/engine/src/main/java/io/aklivity/zilla/runtime/engine/internal/layouts/Layout.java
@@ -15,7 +15,6 @@
  */
 package io.aklivity.zilla.runtime.engine.internal.layouts;
 
-
 public abstract class Layout implements AutoCloseable
 {
     public enum Mode
